supported mimetype and rule refactoring for add-ons
First version of the refactoring, it might need some cleanup and optimization.

Add-on metadata is now stored as a flattened BMessage attached to each add-on
binary and then read when the archiver is loaded.  The format used for the data
in the rdef file is pretty easy to understand but I should document it at some
point.

The rules in the default _bzr_rules.txt file have been completely commented out
as it is now only used to extend the list of rules provided by the archivers.