simplify compression level menu handling
Archiver can now store a pointer m_compressionMenu since most of the archivers
need one.  This allows us avoid tracking and retranslating strings when doing a
BMenu::FindItem().

Also, added Archiver::GetCompressionLevel() which will parse a standard
compression menu that has a numeric prefix and return it as an int32.