addon settings refactoring
Archiver now has complete control over its menu.  Archiver::BuildDefaultMenu()
has been replaced with Archiver::BuildMenu(BMessage&).  The message is either
empty for a default menu or contains any stored settings that the
add-on needs while building the menu.

Archiver now has a ArchiveSettings(BMessage&) method in which each add-on can
store any variables it needs.  The message will be passed to BuildMenu() when
the add-on is created by MainWindow.

The messages are now using untranslated strings for storing settings but we may
want to switch to some other constant which is less volatile than user
interface strings.