PrefsView layout adjustments (related to issue #13)
Attempt to tighten up the spacing on certain preference views in order to
shrink the window back down.

Added some struts in certain areas to show that the controls are related to a
category.

The "Extract path" radio group in PrefsViewPaths was replaced with a checkbox.
I'm still debating whether it's better or not.  Have to check and see if
Teknomancer has an opinion on it.