ArchiverMgr: promote to C++ class
RuleMgr will probably merged into this class or be kept as a member of it.
Eventually this class should do some tracking of loaded add-ons so we can
avoid calling load_add_on()/get_image_symbol() if it's already loaded.
BeOS/Haiku don't really have a problem with loading an add-on multiple times
but it's not necessary.